requirements.txt
requests>=2.31.0
PySocks>=1.7.1 